move the launcher code to a main.py
